fix(snapshots): Hide diff overlay while image loads and deduplicate computeMaskSize
LazyImage rendered children unconditionally, showing the DiffOverlay
on top of the placeholder before the actual image loaded. Now children
only render after the image has loaded.

Also extracts the duplicated computeMaskSize function from both
diffImageDisplay.tsx and snapshotDiffBodies.tsx into a shared module.
