+++ static/app/views/preprod/snapshots/main/snapshotDiffBodies.tsx
@@ -13,19 +13,9 @@ import {getImageName} from 'sentry/views/preprod/types/snapshotTypes';
 
 import {useSyncedD3Zoom} from './imageDisplay/useD3Zoom';
 import {ZoomControls, zoomTransformStyle} from './imageDisplay/zoomControls';
+import {computeMaskSize} from './computeMaskSize';
 
 export const MAX_IMAGE_HEIGHT = 480;
-
-function computeMaskSize(baseImage: SnapshotImage, headImage: SnapshotImage): string {
-  const headW = headImage.width;
-  const headH = headImage.height;
-  if (!headW || !headH) {
-    return '100% 100%';
-  }
-  const maskW = Math.max(baseImage.width || headW, headW);
-  const maskH = Math.max(baseImage.height || headH, headH);
-  return `${(maskW / headW) * 100}% ${(maskH / headH) * 100}%`;
-}
 
 export const SplitPairBody = memo(function SplitPairBody({
   baseUrl,
@@ -328,7 +318,7 @@ function LazyImage({
         onError={onError}
         style={loaded ? undefined : {position: 'absolute', top: 0, left: 0, opacity: 0}}
       />
-      {children}
+      {loaded && children}
     </Container>
   );
 }
